+++ src/screens/Harbor/WithdrawRequests/WithdrawRequests.style.tsx
@@ -10,19 +10,9 @@ export const styles = StyleSheet.create({
     justifyContent: 'center'
   },
   main: { flex: 1 },
-  listMain: {
-    backgroundColor: COLORS.neutral0,
-    borderRadius: 16,
-    padding: scale(8)
-  },
   listContainer: {
     paddingTop: scale(12),
     backgroundColor: COLORS.neutral100,
     paddingHorizontal: scale(12)
-  },
-  dateWrapper: {
-    padding: scale(8),
-    backgroundColor: COLORS.neutral50,
-    borderRadius: 8
   }
 });
